Lecture #17-PrimeNumberCalculatorFrame2 removed unnecessary boxing/unboxing while settings progress

+++ 17_Swing2/src/main/java/hr/fer/oop/swing2/example04/PrimeNumberCalculatorFrame2.java
@@ -71,7 +71,7 @@ public class PrimeNumberCalculatorFrame2 extends JFrame {
                     timer = new Timer(DELAY, new ActionListener() {
                         @Override
                         public void actionPerformed(ActionEvent e) {
-                            progressBar.setValue((Integer) pnTask.getProgress());
+                            progressBar.setValue(pnTask.getProgress());
                             
                             //stop the timer, when finished with the task
                             if (pnTask.getProgress() == 100) {
